move init to trace context
