+++ Tests/SentryTests/Integrations/Performance/Network/SentryNetworkTrackerTests.swift
@@ -590,7 +590,7 @@ class SentryNetworkTrackerTests: XCTestCase {
         sut.urlSessionTaskResume(task)
 
         let expectedTraceHeader = SentrySDK.currentHub().scope.propagationContext.traceHeader.value()
-        let traceContext = SentrySDK.currentHub().scope.propagationContext.getTraceContext(options: self.fixture.options, userSegment: self.fixture.scope.userObject?.segment)
+        let traceContext = SentryTraceContext(trace: SentrySDK.currentHub().scope.propagationContext.traceId, options: self.fixture.options, userSegment: self.fixture.scope.userObject?.segment)
         let expectedBaggageHeader = traceContext.toBaggage().toHTTPHeader()
         XCTAssertEqual(task.currentRequest?.allHTTPHeaderFields?["baggage"] ?? "", expectedBaggageHeader)
         XCTAssertEqual(task.currentRequest?.allHTTPHeaderFields?["sentry-trace"] ?? "", expectedTraceHeader)
@@ -602,7 +602,7 @@ class SentryNetworkTrackerTests: XCTestCase {
         sut.urlSessionTaskResume(task)
 
         let expectedTraceHeader = SentrySDK.currentHub().scope.propagationContext.traceHeader.value()
-        let traceContext = SentrySDK.currentHub().scope.propagationContext.getTraceContext(options: self.fixture.options, userSegment: self.fixture.scope.userObject?.segment)
+        let traceContext = SentryTraceContext(trace: SentrySDK.currentHub().scope.propagationContext.traceId, options: self.fixture.options, userSegment: self.fixture.scope.userObject?.segment)
         let expectedBaggageHeader = traceContext.toBaggage().toHTTPHeader()
         XCTAssertEqual(task.currentRequest?.allHTTPHeaderFields?["baggage"] ?? "", expectedBaggageHeader)
         XCTAssertEqual(task.currentRequest?.allHTTPHeaderFields?["sentry-trace"] ?? "", expectedTraceHeader)
